+++ EncounterAlerts/MidnightS1/Beloren.lua
@@ -30,7 +30,13 @@ NSI.InitializeAlerts[encID] = function(self)
         difficulties = {14, 15, 16},
         customIcon = 1241162,
         timers = nil,
-        overrides = {pinned = true, BlockCopy = true, Size = 100, Anchor = "TOPLEFT", relativeTo = "TOPLEFT", xOffset = 300, yOffset = -300},
+        pinned = true,
+        BlockCopy = true,
+        Size = 100,
+        Anchor = "TOPLEFT",
+        relativeTo = "TOPLEFT",
+        xOffset = 300,
+        yOffset = -300,
         Preview = FeatherColorIconPreview,
         extraOptions = {
             {
@@ -83,14 +89,15 @@ NSI.InitializeAlerts[encID] = function(self)
         customIcon = 1242792,
         difficulties = {14, 15, 16},
         timers = nil,
-        overrides = {pinned = true, BlockCopy = true},
+        pinned = true,
+        BlockCopy = true,
         HideTimer = true,
         Preview = ColorSwapPreview,
     }
     self:AddEncounterAlert(data)
 
     local data = {group = {nil, "Beloren P1", "Beloren P2"}, internalID = "Gateway", text = "Gateway", DisplayType = "Bar", encID = encID, phase = 1, TTS = true, TTSTimer = 4, dur = 6.6, spellID = 311699,
-    timers = {
+        timers = {
             [15] = {{}, {6.6}, {6.6}},
             [16] = {{}, {6.6}, {6.6}},
         },
@@ -98,21 +105,21 @@ NSI.InitializeAlerts[encID] = function(self)
     self:AddEncounterAlert(data)
 
     local data = {group = {nil, "Beloren P1", "Beloren P2"}, internalID = "Next Hit", text = "Next Hit", DisplayType = "Bar", encID = encID, phase = 1, TTS = false, dur = 3.5, spellID = 1242792,
-    timers = {
+        timers = {
             [16] = {{}, {11.7, 15.2, 18.7, 22.2, 25.7, 29.2, 32.7, 36.2, 39.7, 43.2, 46.7}, {11.7, 15.2, 18.7, 22.2, 25.7, 29.2, 32.7, 36.2, 39.7, 43.2, 46.7}},
         },
     }
     self:AddEncounterAlert(data)
 
     local data = {group = {"Beloren P1", "Beloren P2"}, internalID = "Soaks", text = "Soaks", DisplayType = "Text", encID = encID, phase = 1, TTS = false, dur = 8, spellID = nil,
-    timers = {
+        timers = {
             [16] = {{18.8, 68.8}, {70.6, 120.6, 170.6}},
         },
     }
     self:AddEncounterAlert(data)
 
     local data = {group = {"Beloren P1", "Beloren P2"}, internalID = "Quills", text = "Quills", DisplayType = "Text", encID = encID, phase = 1, TTS = false, dur = 6, spellID = nil,
-    timers = {
+        timers = {
             [16] = {{27.4, 37.4, 47.4, 77.4, 87.4, 97.4}, {79.2, 89.2, 99.2, 129.2, 139.2, 149.2, 179.2}},
         },
     }
@@ -128,7 +135,7 @@ local detectedDurations = { -- Death Drop
 NSI.DetectPhaseChange[encID] = function(self, e, info)
     local now = GetTime()
     if e == "ENCOUNTER_TIMELINE_EVENT_REMOVED" or (not info) or (not self.PhaseSwapTime) or (not (now > self.PhaseSwapTime + 5)) or (not self.EncounterID) or (not self.Phase) then return end
-    local difficultyID = select(3, GetInstanceInfo()) or 0
+    local difficultyID = self:DifficultyCheck({14, 15, 16, 233})
     if not difficultyID or not detectedDurations[difficultyID] then return end
     table.insert(self.Timelines, now)
     if self.Phase >= 2 and ApproximatelyEqual(info.duration, 40, 0.2) then
@@ -156,7 +163,7 @@ NSI.DetectPhaseChange[encID] = function(self, e, info)
 end
 
 NSI.EncounterAlertStart[encID] = function(self, id, preview) -- on ENCOUNTER_START
-    id = id or self:DifficultyCheck(14) or 0
+    id = id or self:DifficultyCheck({14, 15, 16}) or 0
     local featherColor = NSRT.EncounterAlerts[encID][id] and NSRT.EncounterAlerts[encID][id]["Feather Color"]
     local colorSwap = NSRT.EncounterAlerts[encID][id] and NSRT.EncounterAlerts[encID][id]["Color Swap"]