+++ EventHandler.lua
@@ -100,9 +100,8 @@ function NSI:EventHandler(e, wowevent, internal, ...) -- internal checks whether
     elseif e == "PLAYER_ENTERING_WORLD" then
         local IsLogin, IsReload = ...
         C_Timer.After(0.01, function()
-            local diff = select(3, GetInstanceInfo()) or 0
-            local ForceHide = diff > 17 or diff < 14
-            if ForceHide then self:HideAllReminders(true) end
+            local diff = self:DifficultyCheck({14, 15, 16})
+            if not diff then self:HideAllReminders(true) end
             if self.LoadedProfile then
                 self:UpdateNoteFrame("ReminderFrame", NSRT.ReminderSettings.ReminderFrame, "skip")
                 self:UpdateNoteFrame("PersonalReminderFrame", NSRT.ReminderSettings.PersonalReminderFrame, "skip")
@@ -110,10 +109,10 @@ function NSI:EventHandler(e, wowevent, internal, ...) -- internal checks whether
             end
         end)
     elseif e == "ENCOUNTER_START" and wowevent then -- allow sending fake encounter_start if in debug mode, only send spec info in mythic, heroic and normal raids
-        local diff = select(3, GetInstanceInfo()) or 0
+        local diff = self:DifficultyCheck({14, 15, 16, 220})
         if internal then diff = 16 end
         if not internal then self:LogTimeline(e, ...) end
-        if (diff < 14 or diff > 17) and diff ~= 220 and not NSRT.Settings["Debug"] then return end -- everything else is enabled in lfr, normal, heroic, mythic and story mode because people like to test in there.
+        if not diff then return end -- everything else is enabled in lfr, normal, heroic, mythic and story mode because people like to test in there.
         self.NSRTFrame.generic_display:Hide()
         self.EncounterID = ...
         self:LoadPersReminder(self.EncounterID)
@@ -165,10 +164,10 @@ function NSI:EventHandler(e, wowevent, internal, ...) -- internal checks whether
     elseif e == "ENCOUNTER_END" and wowevent then
         self:LogTimeline(e, ...)
         local encID, encounterName, _, _, kill = ...
-        local diff = select(3, GetInstanceInfo()) or 0
-        if internal then diff = 16 end
+        local diff = self:DifficultyCheck({14, 15, 16, 220})
+        if internal or diff == 233 then diff = 16 end
         self.CustomEvents = {}
-        if (diff < 14 or diff > 17) and diff ~= 220 then return end
+        if not diff then return end
         self:EncounterRegister(nil, nil, nil, nil, true)
         self:InitPrivateAuras()
         self:HideAllReminders(true)
@@ -218,8 +217,7 @@ function NSI:EventHandler(e, wowevent, internal, ...) -- internal checks whether
         end
     elseif e == "READY_CHECK" and wowevent then
         self.ProcessDone = false
-        local diff= select(3, GetInstanceInfo()) or 0
-        if self:DifficultyCheck(14) or diff == 23 then
+        if self:DifficultyCheck({14, 15, 16, 23}) then
             C_Timer.After(1, function()
                 self:EventHandler("NSI_READY_CHECK", false, true)
             end)
@@ -236,12 +234,12 @@ function NSI:EventHandler(e, wowevent, internal, ...) -- internal checks whether
         if C_ChatInfo.InChatMessagingLockdown() then return end
         self.LastBroadcast = GetTime()
         if self:Restricted() then return end
-        if NSRT.Settings["CheckCooldowns"] and self:DifficultyCheck(15) and UnitInRaid("player") then -- only heroic& mythic because in normal you just wanna go fast and don't care about someone having a cd
+        if NSRT.Settings["CheckCooldowns"] and self:DifficultyCheck({15, 16}) and UnitInRaid("player") then -- only heroic& mythic because in normal you just wanna go fast and don't care about someone having a cd
             self:CheckCooldowns()
         end
     elseif e == "NSI_REM_SHARE"  and internal then
         local unit, reminderstring, assigntable, skipcheck = ...
-        if (UnitIsGroupLeader(unit) or (UnitIsGroupAssistant(unit) and skipcheck)) and (self:DifficultyCheck(14) or skipcheck) then -- skipcheck allows manually sent reminders to bypass difficulty checks
+        if (UnitIsGroupLeader(unit) or (UnitIsGroupAssistant(unit) and skipcheck)) and (self:DifficultyCheck({14, 15, 16}) or skipcheck) then -- skipcheck allows manually sent reminders to bypass difficulty checks
             if reminderstring and type(reminderstring) == "string" and reminderstring ~= "" and ((not NSRT.ReminderSettings.OnlyReceiveGuild) or self:IsInSameGuild(unit)) then
                 self.Reminder = reminderstring
                 NSRT.StoredSharedReminder = reminderstring
@@ -259,7 +257,6 @@ function NSI:EventHandler(e, wowevent, internal, ...) -- internal checks whether
             self:ProcessReminder()
             self:UpdateReminderFrame(true)
         end
-        local diff = select(3, GetInstanceInfo()) or 0
         local text = ""
         if UnitLevel("player") < 90 then return end
         self:CheckRaidBuff()
@@ -348,9 +345,9 @@ function NSI:EventHandler(e, wowevent, internal, ...) -- internal checks whether
                 end)
             end
         end
-        if not self:DifficultyCheck(14) then return end
+        if not self:DifficultyCheck({14, 15, 16}) then return end
     elseif e == "ENCOUNTER_TIMELINE_EVENT_ADDED" and wowevent then
-        if not self:DifficultyCheck(14) then return end
+        if not self:DifficultyCheck({14, 15, 16}) then return end
         local info = ...
         if info.source ~= Enum.EncounterTimelineEventSource.Encounter then
             self.CustomEvents = self.CustomEvents or {}
@@ -360,7 +357,7 @@ function NSI:EventHandler(e, wowevent, internal, ...) -- internal checks whether
         self:LogTimeline(e, ...)
         if self:Restricted() and self.EncounterID and self.DetectPhaseChange[self.EncounterID] then self.DetectPhaseChange[self.EncounterID](self, e, info) end
     elseif e == "ENCOUNTER_TIMELINE_EVENT_REMOVED" and wowevent then
-        if not self:DifficultyCheck(14) then return end
+        if not self:DifficultyCheck({14, 15, 16}) then return end
         local eventID = ...
         if self.CustomEvents and self.CustomEvents[eventID] then
             return
@@ -369,7 +366,7 @@ function NSI:EventHandler(e, wowevent, internal, ...) -- internal checks whether
         if self:Restricted() and self.EncounterID and self.DetectPhaseChange[self.EncounterID] then self.DetectPhaseChange[self.EncounterID](self, e, info) end
     elseif e == "ENCOUNTER_TIMELINE_EVENT_STATE_CHANGED" and wowevent then
         local eventID = ...
-        if not self:DifficultyCheck(14) then return end
+        if not self:DifficultyCheck({14, 15, 16}) then return end
         if self.CustomEvents and self.CustomEvents[eventID] then
             return
         end