+++ Media.lua
@@ -51,6 +51,8 @@ NSI.LSM:Register("sound","|cFF4BAAC8Shroom|r", [[Interface\Addons\NorthernSkyRai
 NSI.LSM:Register("sound","|cFF4BAAC8Fung|r", [[Interface\Addons\NorthernSkyRaidTools\Media\Sounds\Fung.ogg]])
 NSI.LSM:Register("sound","|cFF4BAAC8Right|r", [[Interface\Addons\NorthernSkyRaidTools\Media\Sounds\Right.ogg]])
 NSI.LSM:Register("sound","|cFF4BAAC8Left|r", [[Interface\Addons\NorthernSkyRaidTools\Media\Sounds\Left.ogg]])
+NSI.LSM:Register("sound","|cFF4BAAC8Ranged|r", [[Interface\Addons\NorthernSkyRaidTools\Media\Sounds\Ranged.ogg]])
+NSI.LSM:Register("sound","|cFF4BAAC8Boss|r", [[Interface\Addons\NorthernSkyRaidTools\Media\Sounds\Boss.ogg]])
 --Fonts
 NSI.LSM:Register("font","Expressway", [[Interface\Addons\NorthernSkyRaidTools\Media\Fonts\Expressway.TTF]])
 --StatusBars